requirements.txt
# Core dependencies for PythonWorkflow Framework
pandas>=1.3.0
numpy>=1.20.0
pyyaml>=5.4.0

# Optional dependencies for enhanced functionality
openpyxl>=3.0.0  # For Excel export support
